Move CLI components into app directory
Move all components that make up the command line interface into a
separate app directory in the project root and modify the corresponding
import statements.
