+++ app/cli/__init__.py
@@ -1,1 +1,2 @@
 from .subcommands import add_subparsers, configure_logging
+from .ui import CursesUI, CursesUIError
