+++ app/cli/spec_view.py
