--- spiff_example/curses_ui/__init__.py
@@ -1,1 +1,0 @@
-from .ui import CursesUI, CursesUIError
