+++ spiff_example/spiff/curses_handlers.py
@@ -3,8 +3,8 @@ import logging
 
 from jinja2 import Template
 
-from ..curses_ui.user_input import SimpleField, Option, JsonField
-from ..curses_ui.human_task_handler import TaskHandler
+from app.cli.user_input import SimpleField, Option, JsonField
+from app.cli.human_task_handler import TaskHandler
 
 forms_dir = 'bpmn/tutorial/forms'
